Origin: 10f76e592175e5640965f709f3f36d0cacdeb76d
This is an automated commit.

+++ stormancer/src/P2P/RelayConnection.cpp
@@ -16,8 +16,8 @@ namespace Stormancer
 		_serverConnection->send([=](obytestream* s) {
 			(*s) << (byte)MessageIDTypes::ID_P2P_RELAY;
 			(*s) << id();
-			(*s) << priority;
-			(*s) << reliability;
+			(*s) << (uint8)priority;
+			(*s) << (uint8)reliability;
 			if (writer)
 			{
 				writer(s);
